attempting to add about by adding new component and moving uploads out of index.js
