--- src/assignments/project-1/canvas.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-	<app id="can"></app>
-</body>
-</html>
