+++ src/assignments/project-1/components/about.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react'
+
+
+class About extends Component {
+
+	render() {
+
+		return (
+			<div>
+			<header>
+			<a href="">Album</a></header>
+				<h1>About the Bauhaus Postcard Photo Album</h1>
+				<p>Inspired by a visit to <a href="https://www.harvardartmuseums.org/visit/exhibitions/5615/the-bauhaus-and-harvard">the Bauhaus and Harvard Exhibit at Harvard Art Museums.</a>
+				</p>
+				<p>To create a postcard, visit the <a href="http://emilyc-b.com:8084/">draft postcard generator.</a></p>
+			</div>
+		);
+	}
+}
+
+export default About
