+++ src/assignments/project-1/components/card.js
@@ -1,6 +1,7 @@
 
 import React, { Component } from 'react'
 
+
 class Card extends Component {
 
 	render() {
